refactor: Split cmd/authctl/user/user_test.go
